The repeated get_authentication now defined as a local function
