+++ docs/examples/example_projectq_entangle.py
@@ -2,38 +2,19 @@
 This example is copied from https://github.com/ProjectQ-Framework/ProjectQ
 and is covered under the Apache 2.0 license.
 """
-
 import os
-from getpass import getpass
 
 from projectq import MainEngine
 from projectq.backends import ResourceCounter
 from projectq.ops import CNOT, H, Measure, All
 from projectq.setups import restrictedgateset
 
-from quantuminspire.credentials import load_account, get_token_authentication, get_basic_authentication
 from quantuminspire.api import QuantumInspireAPI
 from quantuminspire.projectq.backend_qx import QIBackend
 
-QI_EMAIL = os.getenv('QI_EMAIL')
-QI_PASSWORD = os.getenv('QI_PASSWORD')
+from tools.credentials import get_authentication
+
 QI_URL = os.getenv('API_URL', 'https://api.quantum-inspire.com/')
-
-
-def get_authentication():
-    """ Gets the authentication for connecting to the Quantum Inspire API."""
-    token = load_account()
-    if token is not None:
-        return get_token_authentication(token)
-    else:
-        if QI_EMAIL is None or QI_PASSWORD is None:
-            print('Enter email:')
-            email = input()
-            print('Enter password')
-            password = getpass()
-        else:
-            email, password = QI_EMAIL, QI_PASSWORD
-        return get_basic_authentication(email, password)
 
 
 project_name = 'ProjectQ-entangle'
