+++ docs/examples/example_projectq_grover.py
@@ -4,7 +4,6 @@ and is covered under the Apache 2.0 license.
 """
 import os
 import math
-from getpass import getpass
 
 from projectq import MainEngine
 from projectq.backends import ResourceCounter, Simulator
@@ -12,12 +11,11 @@ from projectq.meta import Compute, Control, Loop, Uncompute
 from projectq.ops import CNOT, CZ, All, H, Measure, X, Z
 from projectq.setups import restrictedgateset
 
-from quantuminspire.credentials import load_account, get_token_authentication, get_basic_authentication
 from quantuminspire.api import QuantumInspireAPI
 from quantuminspire.projectq.backend_qx import QIBackend
 
-QI_EMAIL = os.getenv('QI_EMAIL')
-QI_PASSWORD = os.getenv('QI_PASSWORD')
+from tools.credentials import get_authentication
+
 QI_URL = os.getenv('API_URL', 'https://api.quantum-inspire.com/')
 
 
@@ -88,22 +86,6 @@ def alternating_bits_oracle(eng, qubits, output):
     with Control(eng, qubits):
         X | output
     Uncompute(eng)
-
-
-def get_authentication():
-    """ Gets the authentication for connecting to the Quantum Inspire API."""
-    token = load_account()
-    if token is not None:
-        return get_token_authentication(token)
-    else:
-        if QI_EMAIL is None or QI_PASSWORD is None:
-            print('Enter email:')
-            email = input()
-            print('Enter password')
-            password = getpass()
-        else:
-            email, password = QI_EMAIL, QI_PASSWORD
-        return get_basic_authentication(email, password)
 
 
 # Remote Quantum-Inspire backend #
