+++ docs/examples/example_qiskit_conditional.py
@@ -12,8 +12,6 @@ provides a Quantum Inspire backend that is used to execute the circuit.
 Copyright 2018-19 QuTech Delft. Licensed under the Apache License, Version 2.0.
 """
 import os
-from getpass import getpass
-from quantuminspire.credentials import load_account, get_token_authentication, get_basic_authentication
 
 from qiskit import BasicAer
 from qiskit.circuit import QuantumRegister, ClassicalRegister, QuantumCircuit
@@ -21,25 +19,9 @@ from qiskit import execute
 
 from quantuminspire.qiskit import QI
 
-QI_EMAIL = os.getenv('QI_EMAIL')
-QI_PASSWORD = os.getenv('QI_PASSWORD')
+from tools.credentials import get_authentication
+
 QI_URL = os.getenv('API_URL', 'https://api.quantum-inspire.com/')
-
-
-def get_authentication():
-    """ Gets the authentication for connecting to the Quantum Inspire API."""
-    token = load_account()
-    if token is not None:
-        return get_token_authentication(token)
-    else:
-        if QI_EMAIL is None or QI_PASSWORD is None:
-            print('Enter email:')
-            email = input()
-            print('Enter password')
-            password = getpass()
-        else:
-            email, password = QI_EMAIL, QI_PASSWORD
-        return get_basic_authentication(email, password)
 
 
 authentication = get_authentication()
