+++ docs/examples/sat_example/example_main.py
@@ -1,28 +1,12 @@
-from src.run import generate_sat_qasm, execute_sat_qasm
-from quantuminspire.api import QuantumInspireAPI
-from quantuminspire.credentials import load_account, get_token_authentication, get_basic_authentication
 import os
-from getpass import getpass
 
-QI_EMAIL = os.getenv('QI_EMAIL')
-QI_PASSWORD = os.getenv('QI_PASSWORD')
+from quantuminspire.api import QuantumInspireAPI
+
+from tools.credentials import get_authentication
+
+from src.run import generate_sat_qasm, execute_sat_qasm
+
 QI_URL = os.getenv('API_URL', 'https://api.quantum-inspire.com/')
-
-
-def get_authentication():
-    """ Gets the authentication for connecting to the Quantum Inspire API."""
-    token = load_account()
-    if token is not None:
-        return get_token_authentication(token)
-    else:
-        if QI_EMAIL is None or QI_PASSWORD is None:
-            print('Enter email:')
-            email = input()
-            print('Enter password')
-            password = getpass()
-        else:
-            email, password = QI_EMAIL, QI_PASSWORD
-        return get_basic_authentication(email, password)
 
 
 # Authenticate, create project and define backend
