+++ setup.py
@@ -46,7 +46,7 @@ setup(name='quantum-inspire-examples',
       version=get_version_number(),
       author='QuantumInspire',
       python_requires='>=3.7',
-      package_dir={},
+      package_dir={'': 'src'},
       data_files=[],
       classifiers=[
           'Development Status :: 4 - Beta',
@@ -56,7 +56,7 @@ setup(name='quantum-inspire-examples',
           'Programming Language :: Python :: 3.9',
           'License :: OSI Approved :: Apache Software License'],
       license='Apache 2.0',
-      packages=[''],
+      packages=['tools'],
       install_requires=['boolean.py', 'jupyter', 'matplotlib>=2.1', 'numpy>=1.17', 'projectq>=0.4', 'qiskit>=0.20.0',
                         'nbimporter', 'quantuminspire'],
       extras_require={
